+++ Cargo.toml
@@ -1,15 +1,14 @@
 [package]
 name = "spin-sdk"
-version = "3.1.0-pre0"
-authors = ["Fermyon Engineering <engineering@fermyon.com>"]
-edition = "2021"
-license = "Apache-2.0 WITH LLVM-exception"
-homepage = "https://developer.fermyon.com/spin/v2/rust-components"
-repository = "https://github.com/fermyon/spin-rust-sdk"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
 documentation = "https://docs.rs/spin-sdk"
-rust-version = "1.73"
 readme = "README.md"
-
 description = """
 The Spin Rust SDK makes it easy to build Spin components in Rust.
 """
@@ -21,12 +20,13 @@ name = "spin_sdk"
 anyhow = "1"
 async-trait = "0.1.74"
 form_urlencoded = "1.0"
-spin-macro = { version = "3.1.0-pre0", path = "macro" }
+spin-executor = { version = "3.1.0-pre0", path = "crates/executor" }
+spin-macro = { version = "3.1.0-pre0", path = "crates/macro" }
 thiserror = "1.0.37"
-wit-bindgen = "0.16.0"
+wit-bindgen = { workspace = true }
 routefinder = "0.5.3"
-once_cell = "1.18.0"
-futures = "0.3.28"
+once_cell = { workspace = true }
+futures = { workspace = true }
 bytes = "1"
 hyperium = { package = "http", version = "1.0.0" }
 serde_json = { version = "1.0.96", optional = true }
@@ -59,6 +59,7 @@ members = [
   "examples/wasi-http-streaming-file",
   "test-cases/simple-http",
   "test-cases/simple-redis",
+  "crates/*",
 ]
 
 [dev-dependencies]
@@ -70,3 +71,17 @@ wasmtime = "18.0.1"
 wasmtime-wasi = "18.0.1"
 wasmtime-wasi-http = "18.0.1"
 wit-component = "0.200.0"
+
+[workspace.package]
+version = "3.1.0-pre0"
+authors = ["Fermyon Engineering <engineering@fermyon.com>"]
+edition = "2021"
+license = "Apache-2.0 WITH LLVM-exception"
+repository = "https://github.com/fermyon/spin-rust-sdk"
+rust-version = "1.73"
+homepage = "https://developer.fermyon.com/spin/v2/rust-components"
+
+[workspace.dependencies]
+wit-bindgen = "0.16.0"
+futures = "0.3.28"
+once_cell = "1.18.0"
