+++ crates/executor/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "spin-executor"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = "https://docs.rs/spin-executor"
+readme = "README.md"
+
+description = """
+Spin SDK async executor
+"""
+
+[dependencies]
+once_cell = "1.18.0"
+futures = "0.3.28"
+wit-bindgen = { workspace = true }
