+++ crates/executor/README.md
@@ -1,0 +1,3 @@
+# The spin-executor crate
+
+This crate contains the spin-sdk's async executor.
