+++ crates/macro/wit/deps/cli/command.wit
