+++ crates/macro/wit/deps/cli/environment.wit
