+++ crates/macro/wit/deps/cli/stdio.wit
