+++ crates/macro/wit/deps/clocks/monotonic-clock.wit
