+++ crates/macro/wit/deps/filesystem/preopens.wit
