+++ crates/macro/wit/deps/filesystem/world.wit
