+++ crates/macro/wit/deps/http/handler.wit
