+++ crates/macro/wit/deps/io/error.wit
