+++ crates/macro/wit/deps/io/streams.wit
