+++ crates/macro/wit/deps/sockets/instance-network.wit
