+++ crates/macro/wit/deps/sockets/ip-name-lookup.wit
