+++ crates/macro/wit/deps/sockets/network.wit
