+++ crates/macro/wit/deps/sockets/tcp-create-socket.wit
