+++ crates/macro/wit/deps/sockets/tcp.wit
