+++ crates/macro/wit/deps/sockets/udp.wit
