+++ crates/macro/wit/deps/sockets/world.wit
