+++ crates/macro/wit/inbound-redis.wit
